feat: support decimal quantities and units
Allow fractional order and inventory quantities using BigDecimal, and align backup/DTOs/services/tests.

+++ src/main/java/com/pos/backup/InventoryExport.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,7 +16,7 @@ import java.time.LocalDateTime;
 public class InventoryExport {
     private Long id;
     private Long productId;
-    private int quantity;
+    private BigDecimal quantity;
     private int lowStockThreshold;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime updatedAt;
